Fix norounds errors in reduxWeb and refluxWeb
The observer implementations in reduxWeb and refluxWeb didn't implement
the norounds method, which threw errors in the browser. Interestingly,
this appeared to prevent refluxWeb from rendering at all, but not
reduxWeb.

Different frontends handle norounds differently (some show a dedicated
message, some show nothing), so I added a test for the norounds case
with no assertions--it just ensures that no errors are thrown.

I also modified refluxWeb to update the history after a round is played.
Other frontends do this, although webSpecs doesn't require them to. I
opted not to add coverage for this.