Add Vue frontend
This required making a few changes to webSpecs. Vue updates the dom
asynchronously, so assertions on page content after interacting with the
page have to be jogged out of the event loop. I introduced a
"whenPageUpdates" helper function to keep this readable.

Also, Vue seems to interfere in ways I don't fully understand with
removing elements from the dom. In webSpecs, removing the fixture
element was not removing the content from the previous test, causing
test pollution.

This was resolved by allowing the concrete test implementation to supply
an optional "teardown" function. vueWeb's implementation of webSpecs
supplies a function that invokes the Vue instance's $destroy and
$el.remove methods.